Update python codes to support project level codes (#9)
* update python to support project level codes

* update

* workspace support

* add logs

* some refactoring

1. some refactoring to the python code, so that veval-parameters are not passed around through the parameters of many functions.
2. deleted the corpus folder and its use in utils.ts, as we are not using corpus now

* Update README.md

rank-bm25 is not needed

---------

+++ python/src/veval.py
@@ -250,9 +250,20 @@ class EvalScore:
 
 
 class VEval:
-    def __init__(self, code, logger=None):
+    def __init__(self, code, veval_param = None, logger=None):
         self.logger = logger
         self.code = code
+
+        # These settings are typically needed for multi-file project
+        if veval_param is not None:
+            #For multi-file verification
+            self.main_file, self.submodule, self.extra_args, self.write_file = veval_param
+        else:
+            self.main_file = ""
+            self.submodule = ""
+            self.extra_args = ""
+            self.write_file = ""
+
         # JSON reported by verus, does not include detailed erros(which is reported from rustc)
         self.verus_result = None
         # JSON reported by rustc, including any compliatoin errors and verus verification errors.
@@ -280,22 +291,51 @@ class VEval:
 
     # Run verus on the code and parse the output.
     def eval(self, max_errs=5, json_mode=True, func_name=None) -> None:
-        with tempfile.NamedTemporaryFile(mode="w", delete=False) as f:
+        if self.write_file == "":
+        #Usually we just write the code to be evaluated into a random place
+            code_path = tempfile.NamedTemporaryFile(delete=False).name
+            isTempFile = True
+        else:
+        #In multi-file setting, the code may need to be in a dedicated file to get evaluated
+            isTempFile = False
+            code_path = self.write_file
+
+        with open(code_path, "w") as f:
             f.write(self.code)
-            code_path = f.name
+
+        if self.main_file and self.submodule:
+            #This is a multi-file project to be verified
+            target_path = self.main_file
+        else:
+            target_path = code_path
+
         multiple_errors = f"--multiple-errors {max_errs}" if max_errs > 0 else ""
         err_format = "--output-json --error-format=json" if json_mode else ""
         # cmd = (f"{self.verus_path} {multiple_errors} {err_format} {code_path}").split(" ")
         # Bug fix: code_path may contain white space
-        cmd = (f"{self.verus_path} {multiple_errors} {err_format}").split(" ")
-        cmd += [code_path]
+        cmd = [self.verus_path, target_path]
+        cmd += (f"{multiple_errors} {err_format}").split(" ")
+
+        if self.submodule:
+            cmd += ["--verify-module", self.submodule]
+
         if func_name:
-            cmd += ["--verify-function", func_name, "--verify-root"]
+            cmd += ["--verify-function", func_name]
+
+        if not self.submodule:
+            cmd += ["--verify-root"]
+
+        if self.extra_args:
+            cmd += self.extra_args.split(" ")
+        cmd = [c for c in cmd if c.strip() != ""]
+
         # self.logger.info(f"Running command: {cmd}")
         m = subprocess.run(cmd, capture_output=True, text=True)
         verus_out = m.stdout
         rustc_out = m.stderr
-        os.unlink(code_path)
+
+        if isTempFile:
+            os.unlink(code_path)
 
         self.verus_out = verus_out
         self.rustc_out = rustc_out
